Cross-referencing from docstrings to file_format.md
Implicit references didn't work between .md and .rst, it seems. So I have to include explicit header references. This is outside basic Markdown, but since it lives in the documentation now that's acceptable.
To get my docstrings to recognise the links, I do need to be a little verbose (:ref:`SITET` instead of just SITET), but for what's a footnote to the docstring I think it's work it.
(There might be some way to add custom logic to the parser to automatically see them as links. But that's way too much effort for this project.)