Fallback version of setuptools_scm
This is mainly so we can compile documentation without error.
Is this a good idea? I don't know. It might mask errors.
The alternative is to set an environment variable in the documentation workflow (https://setuptools-scm.readthedocs.io/en/latest/config/#environment-variables) or to fetch the full depth of the git repo (https://stackoverflow.com/a/68959339). But, this way allows the project to compile even if the version number can't be identified (which shouldn't be a fatal error).
I'll just need to confirm that the version detection works when pushing to test pypi/pypi.